requirements.txt
requests
pandas
numpy
openai
python-dotenv
click
